requirements.txt
llama-index
llama-index-embeddings-ollama
llama-index-llms-ollama
streamlit
llama-index-retrievers-bm25
llama-index-vector-stores-qdrant
qdrant_client
ollama
python-dotenv

jupyter
ipywidgets
psutil

pypdf
whisper-cpp-python